Improve create_project wrapper creation
Improved the call made to make_wrapper in the create_project.tcl script
so that it no longer fails when the bd contains locked ip cores. Instead
the script will quietly continue to execute, but the wrapper will not
get created.

This is important for those who run create_project.tcl in newer versions
of vivado, where it is likely the bd will contain locked IP cores.